requirements.txt
django==4.2
djangorestframework==3.14
python-dotenv==1.0
celery==5.3.1
redis==4.5.1
django-celery-beat==2.5.0